==> jupyter/3_save _chroma_db.ipynb <==
import pandas as pd 
import chromadb
import os

chroma_path ="../chroma_db"

df = pd.read_pickle("../data/embedded_data.pkl")



df["id"] = "tweet" + df.index.astype(str)
df.head()

<h1 style="color:orange;"> Creating collection ! ! </h1>



client = chromadb.PersistentClient(path=chroma_path)

try:
    client.delete_collection(name="tweets")
    print("the collectios was deleted")
except:
    print("pass")
    pass

# metadata={"hnsw:space": "cosine"} tells Chroma to use Cosine Similarity for search

collection = client.create_collection(name="tweets", metadata={"hnsw:space": "cosine"})


# We will use a loop (batching) to avoid crashing the computer. We map the DataFrame columns to the 3 ChromaDB parts (ids, embeddings, metadatas)

from tqdm import tqdm

def add_to_collection(collection, df, batch_size=2000):
    total_rows = len(df)
    # Loop from 0 to total_rows in steps of 2000
    embeddings = list()
    for i in tqdm(range(0, total_rows, batch_size),desc=f'add to {collection.name}'):
        
        batch = df.iloc[i : i+ batch_size]
        
        ids = batch['id'].tolist()
        
        embeddings = batch["embedding"].tolist()
        
        metadatas = batch[["airline_sentiment", "cleaned_text"]].to_dict("records")
        
        collection.add(
            ids=ids,
            embeddings=embeddings,
            metadatas=metadatas
        )

print("Starting ingestion for Data...")

add_to_collection(collection, df)
